// src/features/appData/types/appData.ts
export type UpdatedAppDataProps = {
  appId?: string;
  adminId?: string;
  appTitle?: string;
  appSubTitle?: string;
  domain?: string;
  subDomain?: string;
  pixelId?: string;
  accessToken?: string;
  domainApp?: string;
  domainLanding?: string;
  keitaroDomain?: string;
  keitaroCampaign?: string;
  keitaroCampaignId?: string;
  yandexMetrikaId?: string;
  marketerTag?: string;
  oneSignalApiKey?: string;
  oneSignalAppId?: string;
  displayId?: string;
  displayName?: string;
  appStatus?: string; // status
  category?: string;
  origin?: string;
  language?: string;
  headerReviews?: string;
  hundredPlus?: string;
  aboutThisGame?: string;
  updatedDate?: string;
  country?: string;
  casino?: string;
  isExist?: boolean;
  collectionId?: string;
  about?: string;
  isContainsAds?: boolean;
  isInAppPurchases?: boolean;
  isEditorsChoice?: boolean;
  age?: string;
  rating?: string;
  reviewCount?: string;
  version?: string;
  androidVersion?: string;
  createDate?: Date;
  lastUpdate?: Date;
  releaseDate?: Date;
  ageLimit?: string;
  ageRating?: string;
  commentId?: string;
  newFeatures?: string;
  downloadsCount?: string;
};

export type AppDataProps = {
  isExist?: boolean;
  open?: string;
  noPWASupport?: string;
  stayOn?: string;
  add?: string;
  clickIcon?: string;
  pressIcon?: string;
  wait?: string;
  proceed?: string;
  containsAds?: string;
  inAppPurchases?: string;
  headerReviews?: string;
  hundredPlus?: string;
  downloads?: string;
  ageLimit?: string;
  ageRating?: string;
  install?: string;
  wishlist?: string;
  available?: string;
  aboutThisGame?: string;
  updatedOn?: string;
  updatedDate?: Date;
  casino?: string;
  dataSafety?: string;
  safety?: string;
  noInformation?: string;
  seeDetails?: string;
  ratingsAndReviews?: string;
  verified?: string;
  phone?: string;
  tV?: string;
  chromebook?: string;
  tablet?: string;
  reviews?: string;
  fourPointThree?: string;
  fifteenM?: string;
  helpful?: string;
  allReviews?: string;
  whatsNew?: string;
  findHelpful?: string;
  yes?: string;
  no?: string;
  contact?: string;
  newFeatures?: string;
  downloadsCount?: string;
  isContainsAds?: boolean;
  isInAppPurchases?: boolean;
  isEditorsChoice?: boolean;
  age?: number;
  createDate?: Date;
  lastUpdate?: Date;
  releaseDate?: Date;
  rating?: string;
  about?: string;
  androidVersion?: string;
  collectionId: {
    _id?: string;
    adminId?: string;
    name?: string;
    icon?: string;
    screenShots?: string[];
  };
  commentId: {
    _id?: string;
    adminId?: string;
    name?: string;
    reviewObject: [
      {
        name?: string;
        date?: string;
        review?: string;
        helpful?: string;
        helpfulCount?: string;
        photo?: string;
        rating?: string;
        isResponse?: boolean;
        response?: string;
        responseDate?: string;
      }
    ];
    language?: string;
  };
  reviewCount?: string;
  version?: string;
  _id?: string;
  domain?: string;
  subDomain?: string;
  pixelId?: string;
  accessToken?: string;
  domainApp?: string;
  domainLanding?: string;
  keitaroDomain?: string;
  keitaroCampaign?: string;
  keitaroCampaignId?: string;
  yandexMetrikaId?: string;
  oneSignalApiKey?: string;
  oneSignalAppId?: string;
  displayId?: string;
  displayName?: string;
  marketerTag?: string;
  appTitle?: string;
  appSubTitle?: string;
  renderId?: string;
  icon?: string;
  screenShots?: string[];
  reviewObject: [
    {
      name?: string;
      date?: string;
      review?: string;
      helpful?: string;
      helpfulCount?: string;
      photo?: string;
      rating?: string;
      isResponse?: boolean;
      response?: string;
      responseDate?: string;
    }
  ];
};